Drop the grants util and the RLS migration
The util earned none of its three exports. grantSkillsForEffects moves next
to insertEffectWithModifiers, the chokepoint it is called from and the one
move-from-stash already imports; the pure type_specific_data read joins
subtypeGrantsFromEffects in effect-modifiers, which is where those keys are
interpreted. revalidateRevokedSkillGrants was six lines around one
revalidateTag, now inline at the three removal sites. skillRowSupersedes had
a created_at tie-break nothing needs: two granted rows for one skill differ
only by id in every displayed field.

The migration went the wrong way. fighter_effect_skills has no user-facing
writes at all -- nothing updates it, nothing deletes from it (rows go with
the cascade from fighter_effects, which does not consult RLS), and
add_fighter_injury inserts as SECURITY DEFINER. Granting users INSERT to fit
one derived write widens a surface that should stay shut, so the link insert
uses the service role and the policies stay as they are. Nothing about this
feature needs applying by hand now.
